Removed misleading links from homepage
The links to cotainer homepages in the first column of the overview table was misleading users to the wrong page, instead of the link in the last column (Usage documentation  for a particular tested infrastructure). Solution was to remove the homepage link from front page, it's a mandatory field in the template for conatiner specific test/usage page anyways.